# docs/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ActiveKeyLimitReachedProblem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "ACTIVE_KEY_LIMIT_REACHED"
            ],
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              409
            ],
            "type": "integer"
          },
          "title": {
            "enum": [
              "Active API key limit reached"
            ],
            "type": "string"
          },
          "type": {
            "enum": [
              "urn:api-usage-quota-service:problem:active-key-limit-reached"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "requestId"
        ],
        "type": "object"
      },
...
      "ApiKeyCreateResponseModel": {
        "additionalProperties": false,
        "properties": {
          "apiKey": {
            "$ref": "#/components/schemas/ActiveApiKeyMetadataModel"
          },
          "secret": {
            "format": "password",
            "pattern": "^mq_[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}\\.[A-Za-z0-9_-]{43}$",
            "readOnly": true,
            "type": "string"
          }
        },
        "required": [
          "apiKey",
          "secret"
        ],
        "type": "object"
      },
      "ApiKeyMetadataModel": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "prefix": {
            "pattern": "^mq_[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
            "type": "string"
          },
          "revokedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "scopes": {
            "items": {
              "enum": [
                "usage:write",
                "usage:read",
                "keys:manage",
                "audit:read"
              ],
              "type": "string"
            },
            "maxItems": 4,
            "minItems": 1,
            "type": "array",
            "uniqueItems": true
          },
          "status": {
            "enum": [
              "ACTIVE",
              "REVOKED"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "prefix",
          "scopes",
          "status",
          "createdAt",
          "revokedAt"
        ],
        "type": "object"
      },
      "ApiKeyPageModel": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApiKeyMetadataModel"
            },
            "type": "array"
          },
          "nextCursor": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "type": "object"
      },
...
      "CannotRevokeCurrentKeyProblem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "CANNOT_REVOKE_CURRENT_KEY"
            ],
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              409
            ],
            "type": "integer"
          },
          "title": {
            "enum": [
              "Cannot revoke current API key"
            ],
            "type": "string"
          },
          "type": {
            "enum": [
              "urn:api-usage-quota-service:problem:cannot-revoke-current-key"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "requestId"
        ],
        "type": "object"
      },
...
      "CreateApiKeyDto": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^\\S(?:[\\s\\S]*\\S)?$",
            "type": "string"
          },
          "scopes": {
            "items": {
              "enum": [
                "usage:write",
                "usage:read",
                "keys:manage",
                "audit:read"
              ],
              "type": "string"
            },
            "maxItems": 4,
            "minItems": 1,
            "type": "array",
            "uniqueItems": true
          }
        },
        "required": [
          "name",
          "scopes"
        ],
        "type": "object"
      },
...
      "DependencyUnavailableProblem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "DEPENDENCY_UNAVAILABLE"
            ],
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              503
            ],
            "type": "integer"
          },
          "title": {
            "enum": [
              "Dependency unavailable"
            ],
            "type": "string"
          },
          "type": {
            "enum": [
              "urn:api-usage-quota-service:problem:dependency-unavailable"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "requestId"
        ],
        "type": "object"
      },
...
      "InsufficientScopeProblem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "INSUFFICIENT_SCOPE"
            ],
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              403
            ],
            "type": "integer"
          },
          "title": {
            "enum": [
              "Insufficient scope"
            ],
            "type": "string"
          },
          "type": {
            "enum": [
              "urn:api-usage-quota-service:problem:insufficient-scope"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "requestId"
        ],
        "type": "object"
      },
      "InternalErrorProblem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "INTERNAL_ERROR"
            ],
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              500
            ],
            "type": "integer"
          },
          "title": {
            "enum": [
              "Internal server error"
            ],
            "type": "string"
          },
          "type": {
            "enum": [
              "urn:api-usage-quota-service:problem:internal-error"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "requestId"
        ],
        "type": "object"
      },
      "InvalidApiKeyProblem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "INVALID_API_KEY"
            ],
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              401
            ],
            "type": "integer"
          },
          "title": {
            "enum": [
              "Invalid API key"
            ],
            "type": "string"
          },
          "type": {
            "enum": [
              "urn:api-usage-quota-service:problem:invalid-api-key"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "requestId"
        ],
        "type": "object"
      },
      "InvalidCursorProblem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "INVALID_CURSOR"
            ],
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ValidationErrorItem"
            },
            "type": "array"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              400
            ],
            "type": "integer"
          },
          "title": {
            "enum": [
              "Invalid cursor"
            ],
            "type": "string"
          },
          "type": {
            "enum": [
              "urn:api-usage-quota-service:problem:invalid-cursor"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "requestId"
        ],
        "type": "object"
      },
...
      "ResourceNotFoundProblem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "RESOURCE_NOT_FOUND"
            ],
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              404
            ],
            "type": "integer"
          },
          "title": {
            "enum": [
              "Resource not found"
            ],
            "type": "string"
          },
          "type": {
            "enum": [
              "urn:api-usage-quota-service:problem:resource-not-found"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "requestId"
        ],
        "type": "object"
      },
      "UnsupportedMediaTypeProblem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "UNSUPPORTED_MEDIA_TYPE"
            ],
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              415
            ],
            "type": "integer"
          },
          "title": {
            "enum": [
              "Unsupported media type"
            ],
            "type": "string"
          },
          "type": {
            "enum": [
              "urn:api-usage-quota-service:problem:unsupported-media-type"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "requestId"
        ],
        "type": "object"
      },
      "ValidationErrorItem": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "reason"
        ],
        "type": "object"
      },
      "ValidationErrorProblem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "VALIDATION_ERROR"
            ],
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ValidationErrorItem"
            },
            "type": "array"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              400
            ],
            "type": "integer"
          },
          "title": {
            "enum": [
              "Validation failed"
            ],
            "type": "string"
          },
          "type": {
            "enum": [
              "urn:api-usage-quota-service:problem:validation-error"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "requestId"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "metricsBearer": {
        "bearerFormat": "opaque",
        "scheme": "bearer",
        "type": "http"
      },
      "projectApiKeyBearer": {
        "bearerFormat": "opaque",
        "scheme": "bearer",
        "type": "http"
      },
      "systemAdminBearer": {
        "bearerFormat": "opaque",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Tenant-isolated API usage metering and daily quota service.",
    "title": "API Usage Metering and Quota Service",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/api-keys": {
      "get": {
        "operationId": "ApiKeysController.list",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Opaque unpadded base64url pagination cursor.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_-]+$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyPageModel"
                }
              }
            },
            "description": "",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/InvalidCursorProblem"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationErrorProblem"
                    }
                  ]
                }
              }
            },
            "description": "Problem Details (400).",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidApiKeyProblem"
                }
              }
            },
            "description": "Problem Details (401).",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer authentication challenge.",
                "required": true,
                "schema": {
                  "enum": [
                    "Bearer"
                  ],
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InsufficientScopeProblem"
                }
              }
            },
            "description": "Problem Details (403).",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorProblem"
                }
              }
            },
            "description": "Unexpected internal server error.",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/DependencyUnavailableProblem"
                }
              }
            },
            "description": "Required dependency unavailable.",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "projectApiKeyBearer": []
          }
        ],
        "tags": [
          "api-keys"
        ]
      },
      "post": {
        "operationId": "ApiKeysController.create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyCreateResponseModel"
                }
              }
            },
            "description": "",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorProblem"
                }
              }
            },
            "description": "Problem Details (400).",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidApiKeyProblem"
                }
              }
            },
            "description": "Problem Details (401).",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer authentication challenge.",
                "required": true,
                "schema": {
                  "enum": [
                    "Bearer"
                  ],
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InsufficientScopeProblem"
                }
              }
            },
            "description": "Problem Details (403).",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveKeyLimitReachedProblem"
                }
              }
            },
            "description": "Problem Details (409).",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "415": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/UnsupportedMediaTypeProblem"
                }
              }
            },
            "description": "Problem Details (415).",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorProblem"
                }
              }
            },
            "description": "Unexpected internal server error.",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/DependencyUnavailableProblem"
                }
              }
            },
            "description": "Required dependency unavailable.",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "projectApiKeyBearer": []
          }
        ],
        "tags": [
          "api-keys"
        ]
      }
    },
    "/v1/api-keys/{id}": {
      "delete": {
        "operationId": "ApiKeysController.revoke",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorProblem"
                }
              }
            },
            "description": "Problem Details (400).",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidApiKeyProblem"
                }
              }
            },
            "description": "Problem Details (401).",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer authentication challenge.",
                "required": true,
                "schema": {
                  "enum": [
                    "Bearer"
                  ],
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InsufficientScopeProblem"
                }
              }
            },
            "description": "Problem Details (403).",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceNotFoundProblem"
                }
              }
            },
            "description": "Problem Details (404).",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/CannotRevokeCurrentKeyProblem"
                }
              }
            },
            "description": "Problem Details (409).",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorProblem"
                }
              }
            },
            "description": "Unexpected internal server error.",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/DependencyUnavailableProblem"
                }
              }
            },
            "description": "Required dependency unavailable.",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "projectApiKeyBearer": []
          }
        ],
        "tags": [
          "api-keys"
        ]
      }
    },
...
  },
  "servers": [],
...
}